backend/requirements.txt
# Web Framework and Server
fastapi
uvicorn

# Asynchronous Tasks
celery
redis

# Database
sqlalchemy
psycopg2-binary

# Supabase & AWS Clients
supabase
boto3

# AI and Audio Libraries
google-generativeai
elevenlabs
pydub
PyMuPDF

# Utilities
python-dotenv
fastapi-limiter
slowapi
PyJWT
kombu
python-ulid
httpx
